+++ source/icu.net/BiDi/BiDi.cs
@@ -123,7 +123,28 @@ namespace Icu
 		}
 
 
-		private readonly NativeHandle _biDi = new NativeHandle(nameof(BiDi), NativeMethods.ubidi_close);
+		internal sealed class SafeBiDiHandle : SafeIcuHandle
+		{
+			protected override bool ReleaseIcuHandle()
+			{
+				NativeMethods.ubidi_close(handle);
+				return true;
+			}
+		}
+
+		private readonly SafeBiDiHandle _biDi;
+
+		/// <summary>The handle to pass to ICU.</summary>
+		/// <exception cref="ObjectDisposedException">The ICU libraries were unloaded by
+		/// <see cref="Wrapper.Cleanup"/> after this BiDi object was created.</exception>
+		private SafeBiDiHandle Handle
+		{
+			get
+			{
+				_biDi.ThrowIfStale(nameof(BiDi));
+				return _biDi;
+			}
+		}
 
 		// The ICU Bidi object accepts pointers and expects that the caller keeps buffers allocated, so we handle allocating unmananged memory
 		private IntPtr _para;
@@ -134,8 +155,8 @@ namespace Icu
 		/// </summary>
 		public BiDi()
 		{
-			_biDi.Set(NativeMethods.ubidi_open());
-			if (!_biDi.IsOpen)
+			_biDi = NativeMethods.ubidi_open();
+			if (_biDi.IsInvalid)
 				throw new Exception("Creating BiDi object failed!");
 		}
 
@@ -146,17 +167,17 @@ namespace Icu
 		/// <param name="maxRunCount">The maximum anticipated number of same-level runs that internal memory will be preallocated for</param>
 		public BiDi(int maxLength, int maxRunCount)
 		{
-			_biDi.Set(NativeMethods.ubidi_openSized(maxLength, maxRunCount, out var errorCode));
+			_biDi = NativeMethods.ubidi_openSized(maxLength, maxRunCount, out var errorCode);
 			ExceptionFromErrorCode.ThrowIfError(errorCode, "Creating BiDi object failed! " + errorCode);
 		}
 
 		/// <summary>
 		/// Private constructor.
 		/// </summary>
-		/// <param name="bidiPtr"></param>
-		private BiDi(IntPtr bidiPtr)
+		/// <param name="bidi"></param>
+		private BiDi(SafeBiDiHandle bidi)
 		{
-			_biDi.Set(bidiPtr);
+			_biDi = bidi;
 		}
 
 		/// <summary>
@@ -169,11 +190,7 @@ namespace Icu
 		void Dispose(bool disposing)
 		{
 			if (disposing)
-			{
-				// TODO: dispose managed state (managed objects).
-			}
-
-			_biDi.Close();
+				_biDi.Dispose();
 
 			if (_para != IntPtr.Zero)
 			{
@@ -207,12 +224,12 @@ namespace Icu
 		{
 			get
 			{
-				return NativeMethods.ubidi_getReorderingMode(_biDi.Pointer);
+				return NativeMethods.ubidi_getReorderingMode(Handle);
 			}
 
 			set
 			{
-				NativeMethods.ubidi_setReorderingMode(_biDi.Pointer, value);
+				NativeMethods.ubidi_setReorderingMode(Handle, value);
 			}
 		}
 
@@ -223,12 +240,12 @@ namespace Icu
 		{
 			get
 			{
-				return NativeMethods.ubidi_getReorderingOptions(_biDi.Pointer);
+				return NativeMethods.ubidi_getReorderingOptions(Handle);
 			}
 
 			set
 			{
-				NativeMethods.ubidi_setReorderingOptions(_biDi.Pointer, value);
+				NativeMethods.ubidi_setReorderingOptions(Handle, value);
 			}
 		}
 
@@ -239,12 +256,12 @@ namespace Icu
 		{
 			get
 			{
-				return NativeMethods.ubidi_isInverse(_biDi.Pointer);
+				return NativeMethods.ubidi_isInverse(Handle);
 			}
 
 			set
 			{
-				NativeMethods.ubidi_setInverse(_biDi.Pointer, value);
+				NativeMethods.ubidi_setInverse(Handle, value);
 			}
 		}
 
@@ -255,12 +272,12 @@ namespace Icu
 		{
 			get
 			{
-				return NativeMethods.ubidi_isOrderParagraphsLTR(_biDi.Pointer);
+				return NativeMethods.ubidi_isOrderParagraphsLTR(Handle);
 			}
 
 			set
 			{
-				NativeMethods.ubidi_orderParagraphsLTR(_biDi.Pointer, value);
+				NativeMethods.ubidi_orderParagraphsLTR(Handle, value);
 			}
 		}
 
@@ -271,7 +288,7 @@ namespace Icu
 		{
 			get
 			{
-				return NativeMethods.ubidi_getDirection(_biDi.Pointer);
+				return NativeMethods.ubidi_getDirection(Handle);
 			}
 		}
 
@@ -310,7 +327,7 @@ namespace Icu
 
 			// icu BiDi expects the para pointer to live for the life of the structure, so we have to stash it
 			_para = Marshal.StringToHGlobalUni(text);
-			NativeMethods.ubidi_setPara(_biDi.Pointer, _para, text.Length, paraLevel, embeddingLevels, out var errorCode);
+			NativeMethods.ubidi_setPara(Handle, _para, text.Length, paraLevel, embeddingLevels, out var errorCode);
 			ExceptionFromErrorCode.ThrowIfError(errorCode, "BiDi analysis failed! " + errorCode);
 		}
 
@@ -322,7 +339,7 @@ namespace Icu
 		/// <returns></returns>
 		public BiDi SetLine(int start, int limit)
 		{
-			NativeMethods.ubidi_setLine(_biDi.Pointer, start, limit, out var lineBidi, out var errorCode);
+			NativeMethods.ubidi_setLine(Handle, start, limit, out var lineBidi, out var errorCode);
 			ExceptionFromErrorCode.ThrowIfError(errorCode, "BiDi line creation failed! " + errorCode);
 			return new BiDi(lineBidi);
 		}
@@ -334,7 +351,7 @@ namespace Icu
 		{
 			get
 			{
-				return Marshal.PtrToStringUni(NativeMethods.ubidi_getText(_biDi.Pointer), NativeMethods.ubidi_getLength(_biDi.Pointer));
+				return Marshal.PtrToStringUni(NativeMethods.ubidi_getText(Handle), NativeMethods.ubidi_getLength(Handle));
 			}
 		}
 
@@ -345,7 +362,7 @@ namespace Icu
 		{
 			get
 			{
-				return NativeMethods.ubidi_getLength(_biDi.Pointer);
+				return NativeMethods.ubidi_getLength(Handle);
 			}
 		}
 
@@ -356,7 +373,7 @@ namespace Icu
 		{
 			get
 			{
-				return NativeMethods.ubidi_getProcessedLength(_biDi.Pointer);
+				return NativeMethods.ubidi_getProcessedLength(Handle);
 			}
 		}
 
@@ -367,7 +384,7 @@ namespace Icu
 		{
 			get
 			{
-				return NativeMethods.ubidi_getResultLength(_biDi.Pointer);
+				return NativeMethods.ubidi_getResultLength(Handle);
 			}
 		}
 
@@ -377,7 +394,7 @@ namespace Icu
 		/// <returns></returns>
 		public IEnumerable<byte> GetLevels()
 		{
-			var levels = NativeMethods.ubidi_getLevels(_biDi.Pointer, out var errorCode);
+			var levels = NativeMethods.ubidi_getLevels(Handle, out var errorCode);
 			ExceptionFromErrorCode.ThrowIfError(errorCode, "BiDi level retrieval failed! " + errorCode);
 
 			var ret = new byte[ProcessedLength];
@@ -392,7 +409,7 @@ namespace Icu
 		/// <returns></returns>
 		public byte GetLevelAt(int index)
 		{
-			return NativeMethods.ubidi_getLevelAt(_biDi.Pointer, index);
+			return NativeMethods.ubidi_getLevelAt(Handle, index);
 		}
 
 		/// <summary>
@@ -401,7 +418,7 @@ namespace Icu
 		/// <returns>The paragraph level</returns>
 		public byte GetParaLevel()
 		{
-			return NativeMethods.ubidi_getParaLevel(_biDi.Pointer);
+			return NativeMethods.ubidi_getParaLevel(Handle);
 		}
 
 		/// <summary>
@@ -410,7 +427,7 @@ namespace Icu
 		/// <returns>The number of paragraphs</returns>
 		public int CountParagraphs()
 		{
-			return NativeMethods.ubidi_countParagraphs(_biDi.Pointer);
+			return NativeMethods.ubidi_countParagraphs(Handle);
 		}
 
 		/// <summary>
@@ -423,7 +440,7 @@ namespace Icu
 		/// <returns>The index of the paragraph containing the specified position</returns>
 		public int GetParagraph(int charIndex, out int paraStart, out int paraLimit, out byte paraLevel)
 		{
-			var ret = NativeMethods.ubidi_getParagraph(_biDi.Pointer, charIndex, out paraStart, out paraLimit, out paraLevel, out var errorCode);
+			var ret = NativeMethods.ubidi_getParagraph(Handle, charIndex, out paraStart, out paraLimit, out paraLevel, out var errorCode);
 			ExceptionFromErrorCode.ThrowIfError(errorCode, "Paragraph retrieval failed! " + errorCode);
 			return ret;
 		}
@@ -437,7 +454,7 @@ namespace Icu
 		/// <param name="paraLevel">Will receive the level of the paragraph</param>
 		public void GetParagraphByIndex(int paraIndex, out int paraStart, out int paraLimit, out byte paraLevel)
 		{
-			NativeMethods.ubidi_getParagraphByIndex(_biDi.Pointer, paraIndex, out paraStart, out paraLimit, out paraLevel, out var errorCode);
+			NativeMethods.ubidi_getParagraphByIndex(Handle, paraIndex, out paraStart, out paraLimit, out paraLevel, out var errorCode);
 			ExceptionFromErrorCode.ThrowIfError(errorCode, "Paragraph retrieval failed! " + errorCode);
 		}
 
@@ -447,7 +464,7 @@ namespace Icu
 		/// <returns>The number of runs</returns>
 		public int CountRuns()
 		{
-			var ret = NativeMethods.ubidi_countRuns(_biDi.Pointer, out var errorCode);
+			var ret = NativeMethods.ubidi_countRuns(Handle, out var errorCode);
 			ExceptionFromErrorCode.ThrowIfError(errorCode, "Run count failed! " + errorCode);
 			return ret;
 		}
@@ -459,7 +476,7 @@ namespace Icu
 		/// <returns>The index of this character in the text</returns>
 		public int GetLogicalIndex(int visualIndex)
 		{
-			var ret = NativeMethods.ubidi_getLogicalIndex(_biDi.Pointer, visualIndex, out var errorCode);
+			var ret = NativeMethods.ubidi_getLogicalIndex(Handle, visualIndex, out var errorCode);
 			ExceptionFromErrorCode.ThrowIfError(errorCode, "Get logical index failed! " + errorCode);
 			return ret;
 		}
@@ -471,7 +488,7 @@ namespace Icu
 		public int[] GetLogicalMap()
 		{
 			var map = new int[ResultLength];
-			NativeMethods.ubidi_getLogicalMap(_biDi.Pointer, map, out var errorCode);
+			NativeMethods.ubidi_getLogicalMap(Handle, map, out var errorCode);
 			ExceptionFromErrorCode.ThrowIfError(errorCode, "Get logical map failed! " + errorCode);
 			return map;
 		}
@@ -484,7 +501,7 @@ namespace Icu
 		/// <returns>The limit of the corresponding run</returns>
 		public int GetLogicalRun(int logicalPosition, out byte runLevel)
 		{
-			NativeMethods.ubidi_getLogicalRun(_biDi.Pointer, logicalPosition, out int limit, out runLevel);
+			NativeMethods.ubidi_getLogicalRun(Handle, logicalPosition, out int limit, out runLevel);
 			return limit;
 		}
 
@@ -495,7 +512,7 @@ namespace Icu
 		/// <returns>The visual position of this character.</returns>
 		public int GetVisualIndex(int logicalIndex)
 		{
-			var ret = NativeMethods.ubidi_getVisualIndex(_biDi.Pointer, logicalIndex, out var errorCode);
+			var ret = NativeMethods.ubidi_getVisualIndex(Handle, logicalIndex, out var errorCode);
 			ExceptionFromErrorCode.ThrowIfError(errorCode, "Get visual index failed! " + errorCode);
 			return ret;
 		}
@@ -507,7 +524,7 @@ namespace Icu
 		public int[] GetVisualMap()
 		{
 			var map = new int[ResultLength];
-			NativeMethods.ubidi_getVisualMap(_biDi.Pointer, map, out var errorCode);
+			NativeMethods.ubidi_getVisualMap(Handle, map, out var errorCode);
 			ExceptionFromErrorCode.ThrowIfError(errorCode, "Get visual map failed! " + errorCode);
 			return map;
 		}
@@ -521,7 +538,7 @@ namespace Icu
 		/// <returns>The directionality of the run, never UBIDI_MIXED, never UBIDI_NEUTRAL</returns>
 		public BiDiDirection GetVisualRun(int runIndex, out int logicalStart, out int runLength)
 		{
-			return NativeMethods.ubidi_getVisualRun(_biDi.Pointer, runIndex, out logicalStart, out runLength);
+			return NativeMethods.ubidi_getVisualRun(Handle, runIndex, out logicalStart, out runLength);
 		}
 
 		/// <summary>
@@ -533,7 +550,7 @@ namespace Icu
 		public string GetReordered(CallReorderingOptions options)
 		{
 			var buff = new char[ProcessedLength * 2];
-			var len = NativeMethods.ubidi_writeReordered(_biDi.Pointer, buff, buff.Length * 2, (ushort)options, out var errorCode);
+			var len = NativeMethods.ubidi_writeReordered(Handle, buff, buff.Length * 2, (ushort)options, out var errorCode);
 			ExceptionFromErrorCode.ThrowIfError(errorCode, "BiDi reordering failed! " + errorCode);
 
 			return new string(buff, 0, len);