+++ source/icu.net/NativeMethods/NativeMethods_BiDi.cs
@@ -15,100 +15,100 @@ namespace Icu
 		private class BiDiMethodsContainer
 		{
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate IntPtr ubidi_openDelegate();
+			internal delegate BiDi.SafeBiDiHandle ubidi_openDelegate();
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate IntPtr ubidi_openSizedDelegate(int maxLength, int maxRunCount, out ErrorCode errorCode);
+			internal delegate BiDi.SafeBiDiHandle ubidi_openSizedDelegate(int maxLength, int maxRunCount, out ErrorCode errorCode);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 			internal delegate void ubidi_closeDelegate(IntPtr bidi);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate BiDi.BiDiReorderingMode ubidi_getReorderingModeDelegate(IntPtr bidi);
+			internal delegate BiDi.BiDiReorderingMode ubidi_getReorderingModeDelegate(BiDi.SafeBiDiHandle bidi);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate void ubidi_setReorderingModeDelegate(IntPtr bidi, BiDi.BiDiReorderingMode reorderingMode);
+			internal delegate void ubidi_setReorderingModeDelegate(BiDi.SafeBiDiHandle bidi, BiDi.BiDiReorderingMode reorderingMode);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate BiDi.BiDiReorderingOption ubidi_getReorderingOptionsDelegate(IntPtr bidi);
+			internal delegate BiDi.BiDiReorderingOption ubidi_getReorderingOptionsDelegate(BiDi.SafeBiDiHandle bidi);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate void ubidi_setReorderingOptionsDelegate(IntPtr bidi, BiDi.BiDiReorderingOption reorderingOptions);
+			internal delegate void ubidi_setReorderingOptionsDelegate(BiDi.SafeBiDiHandle bidi, BiDi.BiDiReorderingOption reorderingOptions);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate bool ubidi_isInverseDelegate(IntPtr bidi);
+			internal delegate bool ubidi_isInverseDelegate(BiDi.SafeBiDiHandle bidi);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate void ubidi_setInverseDelegate(IntPtr bidi, bool isInverse);
+			internal delegate void ubidi_setInverseDelegate(BiDi.SafeBiDiHandle bidi, bool isInverse);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate bool ubidi_isOrderParagraphsLTRDelegate(IntPtr bidi);
+			internal delegate bool ubidi_isOrderParagraphsLTRDelegate(BiDi.SafeBiDiHandle bidi);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate void ubidi_orderParagraphsLTRDelegate(IntPtr bidi, bool orderParagraphsLTR);
+			internal delegate void ubidi_orderParagraphsLTRDelegate(BiDi.SafeBiDiHandle bidi, bool orderParagraphsLTR);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate BiDi.BiDiDirection ubidi_getDirectionDelegate(IntPtr bidi);
+			internal delegate BiDi.BiDiDirection ubidi_getDirectionDelegate(BiDi.SafeBiDiHandle bidi);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate void ubidi_setParaDelegate(IntPtr bidi, IntPtr text, int length, byte paraLevel, [MarshalAs(UnmanagedType.LPArray)]byte[] embeddingLevels, out ErrorCode errorCode);
+			internal delegate void ubidi_setParaDelegate(BiDi.SafeBiDiHandle bidi, IntPtr text, int length, byte paraLevel, [MarshalAs(UnmanagedType.LPArray)]byte[] embeddingLevels, out ErrorCode errorCode);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate void ubidi_setLineDelegate(IntPtr bidi, int start, int limit, out IntPtr lineBiDi, out ErrorCode errorCode);
+			internal delegate void ubidi_setLineDelegate(BiDi.SafeBiDiHandle bidi, int start, int limit, out BiDi.SafeBiDiHandle lineBiDi, out ErrorCode errorCode);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate int ubidi_getLengthDelegate(IntPtr bidi);
+			internal delegate int ubidi_getLengthDelegate(BiDi.SafeBiDiHandle bidi);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate int ubidi_getProcessedLengthDelegate(IntPtr bidi);
+			internal delegate int ubidi_getProcessedLengthDelegate(BiDi.SafeBiDiHandle bidi);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate int ubidi_getResultLengthDelegate(IntPtr bidi);
+			internal delegate int ubidi_getResultLengthDelegate(BiDi.SafeBiDiHandle bidi);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate IntPtr ubidi_getTextDelegate(IntPtr bidi);
+			internal delegate IntPtr ubidi_getTextDelegate(BiDi.SafeBiDiHandle bidi);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate IntPtr ubidi_getLevelsDelegate(IntPtr bidi, out ErrorCode errorCode);
+			internal delegate IntPtr ubidi_getLevelsDelegate(BiDi.SafeBiDiHandle bidi, out ErrorCode errorCode);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate byte ubidi_getLevelAtDelegate(IntPtr bidi, int charIndex);
+			internal delegate byte ubidi_getLevelAtDelegate(BiDi.SafeBiDiHandle bidi, int charIndex);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate byte ubidi_getParaLevelDelegate(IntPtr bidi);
+			internal delegate byte ubidi_getParaLevelDelegate(BiDi.SafeBiDiHandle bidi);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate int ubidi_countParagraphsDelegate(IntPtr bidi);
+			internal delegate int ubidi_countParagraphsDelegate(BiDi.SafeBiDiHandle bidi);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate int ubidi_getParagraphDelegate(IntPtr bidi, int charIndex, out int paraStart, out int paraLimit, out byte paraLevel, out ErrorCode errorCode);
+			internal delegate int ubidi_getParagraphDelegate(BiDi.SafeBiDiHandle bidi, int charIndex, out int paraStart, out int paraLimit, out byte paraLevel, out ErrorCode errorCode);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate void ubidi_getParagraphByIndexDelegate(IntPtr bidi, int paraIndex, out int paraStart, out int paraLimit, out byte paraLevel, out ErrorCode errorCode);
+			internal delegate void ubidi_getParagraphByIndexDelegate(BiDi.SafeBiDiHandle bidi, int paraIndex, out int paraStart, out int paraLimit, out byte paraLevel, out ErrorCode errorCode);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate int ubidi_countRunsDelegate(IntPtr bidi, out ErrorCode errorCode);
+			internal delegate int ubidi_countRunsDelegate(BiDi.SafeBiDiHandle bidi, out ErrorCode errorCode);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate int ubidi_getLogicalIndexDelegate(IntPtr bidi, int visualIndex, out ErrorCode errorCode);
+			internal delegate int ubidi_getLogicalIndexDelegate(BiDi.SafeBiDiHandle bidi, int visualIndex, out ErrorCode errorCode);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate void ubidi_getLogicalMapDelegate(IntPtr bidi, [Out, MarshalAs(UnmanagedType.LPArray)]int[] indexMap, out ErrorCode errorCode);
+			internal delegate void ubidi_getLogicalMapDelegate(BiDi.SafeBiDiHandle bidi, [Out, MarshalAs(UnmanagedType.LPArray)]int[] indexMap, out ErrorCode errorCode);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate void ubidi_getLogicalRunDelegate(IntPtr bidi, int logicalPosition, out int logicalLimit, out byte level);
+			internal delegate void ubidi_getLogicalRunDelegate(BiDi.SafeBiDiHandle bidi, int logicalPosition, out int logicalLimit, out byte level);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate int ubidi_getVisualIndexDelegate(IntPtr bidi, int logicalIndex, out ErrorCode errorCode);
+			internal delegate int ubidi_getVisualIndexDelegate(BiDi.SafeBiDiHandle bidi, int logicalIndex, out ErrorCode errorCode);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate void ubidi_getVisualMapDelegate(IntPtr bidi, [Out, MarshalAs(UnmanagedType.LPArray)]int[] indexMap, out ErrorCode errorCode);
+			internal delegate void ubidi_getVisualMapDelegate(BiDi.SafeBiDiHandle bidi, [Out, MarshalAs(UnmanagedType.LPArray)]int[] indexMap, out ErrorCode errorCode);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
-			internal delegate BiDi.BiDiDirection ubidi_getVisualRunDelegate(IntPtr bidi, int runIndex, out int logicalStart, out int length);
+			internal delegate BiDi.BiDiDirection ubidi_getVisualRunDelegate(BiDi.SafeBiDiHandle bidi, int runIndex, out int logicalStart, out int length);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl, CharSet = CharSet.Unicode)]
-			internal delegate int ubidi_writeReorderedDelegate(IntPtr bidi, [Out, MarshalAs(UnmanagedType.LPArray)]char[] dest, int destSize, ushort options, out ErrorCode errorCode);
+			internal delegate int ubidi_writeReorderedDelegate(BiDi.SafeBiDiHandle bidi, [Out, MarshalAs(UnmanagedType.LPArray)]char[] dest, int destSize, ushort options, out ErrorCode errorCode);
 
 			[UnmanagedFunctionPointer(CallingConvention.Cdecl, CharSet = CharSet.Unicode)]
 			internal delegate int ubidi_writeReverseDelegate(string src, int srcLength, [Out, MarshalAs(UnmanagedType.LPArray)]char[] dest, int destSize, ushort options, out ErrorCode errorCode);
@@ -159,7 +159,7 @@ namespace Icu
 		/// Create a UBiDi structure.
 		/// </summary>
 		/// <returns>An created empty UBiDi structure.</returns>
-		public static IntPtr ubidi_open()
+		public static BiDi.SafeBiDiHandle ubidi_open()
 		{
 			if (BiDiMethods.ubidi_open == null)
 				BiDiMethods.ubidi_open = GetMethod<BiDiMethodsContainer.ubidi_openDelegate>(IcuCommonLibHandle, "ubidi_open");
@@ -173,7 +173,7 @@ namespace Icu
 		/// <param name="maxRunCount">The maximum anticipated number of same-level runs that internal memory will be preallocated for</param>
 		/// <param name="errorCode">The error code</param>
 		/// <returns></returns>
-		public static IntPtr ubidi_openSized(int maxLength, int maxRunCount, out ErrorCode errorCode)
+		public static BiDi.SafeBiDiHandle ubidi_openSized(int maxLength, int maxRunCount, out ErrorCode errorCode)
 		{
 			if (BiDiMethods.ubidi_openSized == null)
 				BiDiMethods.ubidi_openSized = GetMethod<BiDiMethodsContainer.ubidi_openSizedDelegate>(IcuCommonLibHandle, "ubidi_openSized");
@@ -196,7 +196,7 @@ namespace Icu
 		/// </summary>
 		/// <param name="bidi">The BiDi object</param>
 		/// <returns>A combination of zero or more of the reordering options</returns>
-		public static BiDi.BiDiReorderingMode ubidi_getReorderingMode(IntPtr bidi)
+		public static BiDi.BiDiReorderingMode ubidi_getReorderingMode(BiDi.SafeBiDiHandle bidi)
 		{
 			if (BiDiMethods.ubidi_getReorderingMode == null)
 				BiDiMethods.ubidi_getReorderingMode = GetMethod<BiDiMethodsContainer.ubidi_getReorderingModeDelegate>(IcuCommonLibHandle, "ubidi_getReorderingMode");
@@ -208,7 +208,7 @@ namespace Icu
 		/// </summary>
 		/// <param name="bidi">The BiDi object</param>
 		/// <param name="reorderingMode">A combination of zero or more of the reordering options</param>
-		public static void ubidi_setReorderingMode(IntPtr bidi, BiDi.BiDiReorderingMode reorderingMode)
+		public static void ubidi_setReorderingMode(BiDi.SafeBiDiHandle bidi, BiDi.BiDiReorderingMode reorderingMode)
 		{
 			if (BiDiMethods.ubidi_setReorderingMode == null)
 				BiDiMethods.ubidi_setReorderingMode = GetMethod<BiDiMethodsContainer.ubidi_setReorderingModeDelegate>(IcuCommonLibHandle, "ubidi_setReorderingMode");
@@ -220,7 +220,7 @@ namespace Icu
 		/// </summary>
 		/// <param name="bidi">The BiDi object</param>
 		/// <returns>A combination of zero or more of the reordering options</returns>
-		public static BiDi.BiDiReorderingOption ubidi_getReorderingOptions(IntPtr bidi)
+		public static BiDi.BiDiReorderingOption ubidi_getReorderingOptions(BiDi.SafeBiDiHandle bidi)
 		{
 			if (BiDiMethods.ubidi_getReorderingOptions == null)
 				BiDiMethods.ubidi_getReorderingOptions = GetMethod<BiDiMethodsContainer.ubidi_getReorderingOptionsDelegate>(IcuCommonLibHandle, "ubidi_getReorderingOptions");
@@ -232,7 +232,7 @@ namespace Icu
 		/// </summary>
 		/// <param name="bidi">The BiDi object</param>
 		/// <param name="reorderingOptions">A combination of zero or more of the reordering options</param>
-		public static void ubidi_setReorderingOptions(IntPtr bidi, BiDi.BiDiReorderingOption reorderingOptions)
+		public static void ubidi_setReorderingOptions(BiDi.SafeBiDiHandle bidi, BiDi.BiDiReorderingOption reorderingOptions)
 		{
 			if (BiDiMethods.ubidi_setReorderingOptions == null)
 				BiDiMethods.ubidi_setReorderingOptions = GetMethod<BiDiMethodsContainer.ubidi_setReorderingOptionsDelegate>(IcuCommonLibHandle, "ubidi_setReorderingOptions");
@@ -244,7 +244,7 @@ namespace Icu
 		/// </summary>
 		/// <param name="bidi">The BiDi object</param>
 		/// <returns>True if the object is using the inverse Bidi algorithm</returns>
-		public static bool ubidi_isInverse(IntPtr bidi)
+		public static bool ubidi_isInverse(BiDi.SafeBiDiHandle bidi)
 		{
 			if (BiDiMethods.ubidi_isInverse == null)
 				BiDiMethods.ubidi_isInverse = GetMethod<BiDiMethodsContainer.ubidi_isInverseDelegate>(IcuCommonLibHandle, "ubidi_isInverse");
@@ -256,7 +256,7 @@ namespace Icu
 		/// </summary>
 		/// <param name="bidi">The BiDi object</param>
 		/// <param name="isInverse">True to use the inverse algorithm</param>
-		public static void ubidi_setInverse(IntPtr bidi, bool isInverse)
+		public static void ubidi_setInverse(BiDi.SafeBiDiHandle bidi, bool isInverse)
 		{
 			if (BiDiMethods.ubidi_setInverse == null)
 				BiDiMethods.ubidi_setInverse = GetMethod<BiDiMethodsContainer.ubidi_setInverseDelegate>(IcuCommonLibHandle, "ubidi_setInverse");
@@ -268,7 +268,7 @@ namespace Icu
 		/// </summary>
 		/// <param name="bidi">The BiDi object</param>
 		/// <returns>True if the object is using level 0 for block separators</returns>
-		public static bool ubidi_isOrderParagraphsLTR(IntPtr bidi)
+		public static bool ubidi_isOrderParagraphsLTR(BiDi.SafeBiDiHandle bidi)
 		{
 			if (BiDiMethods.ubidi_isOrderParagraphsLTR == null)
 				BiDiMethods.ubidi_isOrderParagraphsLTR = GetMethod<BiDiMethodsContainer.ubidi_isOrderParagraphsLTRDelegate>(IcuCommonLibHandle, "ubidi_isOrderParagraphsLTR");
@@ -281,7 +281,7 @@ namespace Icu
 		/// <param name="bidi">The BiDi object</param>
 		/// <param name="orderParagraphsLTR">True to use level 0 for block separators</param>
 		// ReSharper disable once InconsistentNaming
-		public static void ubidi_orderParagraphsLTR(IntPtr bidi, bool orderParagraphsLTR)
+		public static void ubidi_orderParagraphsLTR(BiDi.SafeBiDiHandle bidi, bool orderParagraphsLTR)
 		{
 			if (BiDiMethods.ubidi_orderParagraphsLTR == null)
 				BiDiMethods.ubidi_orderParagraphsLTR = GetMethod<BiDiMethodsContainer.ubidi_orderParagraphsLTRDelegate>(IcuCommonLibHandle, "ubidi_orderParagraphsLTR");
@@ -293,7 +293,7 @@ namespace Icu
 		/// </summary>
 		/// <param name="bidi">The BiDi object</param>
 		/// <returns>The directionality of the text</returns>
-		public static BiDi.BiDiDirection ubidi_getDirection(IntPtr bidi)
+		public static BiDi.BiDiDirection ubidi_getDirection(BiDi.SafeBiDiHandle bidi)
 		{
 			if (BiDiMethods.ubidi_getDirection == null)
 				BiDiMethods.ubidi_getDirection = GetMethod<BiDiMethodsContainer.ubidi_getDirectionDelegate>(IcuCommonLibHandle, "ubidi_getDirection");
@@ -309,7 +309,7 @@ namespace Icu
 		/// <param name="paraLevel">The base paragraph level.</param>
 		/// <param name="embeddingLevels">Preset embedding and override levels</param>
 		/// <param name="errorCode">The error code</param>
-		public static void ubidi_setPara(IntPtr bidi, IntPtr text, int length, byte paraLevel, byte[] embeddingLevels, out ErrorCode errorCode)
+		public static void ubidi_setPara(BiDi.SafeBiDiHandle bidi, IntPtr text, int length, byte paraLevel, byte[] embeddingLevels, out ErrorCode errorCode)
 		{
 			if (BiDiMethods.ubidi_setPara == null)
 				BiDiMethods.ubidi_setPara = GetMethod<BiDiMethodsContainer.ubidi_setParaDelegate>(IcuCommonLibHandle, "ubidi_setPara");
@@ -324,7 +324,7 @@ namespace Icu
 		/// <param name="limit">Index just behind the line's last index into the text (its last index +1)</param>
 		/// <param name="lineBidi">The new line BiDi object</param>
 		/// <param name="errorCode">The error code</param>
-		public static void ubidi_setLine(IntPtr bidi, int start, int limit, out IntPtr lineBidi, out ErrorCode errorCode)
+		public static void ubidi_setLine(BiDi.SafeBiDiHandle bidi, int start, int limit, out BiDi.SafeBiDiHandle lineBidi, out ErrorCode errorCode)
 		{
 			if (BiDiMethods.ubidi_setLine == null)
 				BiDiMethods.ubidi_setLine = GetMethod<BiDiMethodsContainer.ubidi_setLineDelegate>(IcuCommonLibHandle, "ubidi_setLine");
@@ -336,7 +336,7 @@ namespace Icu
 		/// </summary>
 		/// <param name="bidi">The BiDi object</param>
 		/// <returns>The text length</returns>
-		public static int ubidi_getLength(IntPtr bidi)
+		public static int ubidi_getLength(BiDi.SafeBiDiHandle bidi)
 		{
 			if (BiDiMethods.ubidi_getLength == null)
 				BiDiMethods.ubidi_getLength = GetMethod<BiDiMethodsContainer.ubidi_getLengthDelegate>(IcuCommonLibHandle, "ubidi_getLength");
@@ -348,7 +348,7 @@ namespace Icu
 		/// </summary>
 		/// <param name="bidi">The BiDi object</param>
 		/// <returns>The source text length</returns>
-		public static int ubidi_getProcessedLength(IntPtr bidi)
+		public static int ubidi_getProcessedLength(BiDi.SafeBiDiHandle bidi)
 		{
 			if (BiDiMethods.ubidi_getProcessedLength == null)
 				BiDiMethods.ubidi_getProcessedLength = GetMethod<BiDiMethodsContainer.ubidi_getProcessedLengthDelegate>(IcuCommonLibHandle, "ubidi_getProcessedLength");
@@ -360,7 +360,7 @@ namespace Icu
 		/// </summary>
 		/// <param name="bidi">The BiDi object</param>
 		/// <returns>The result text length</returns>
-		public static int ubidi_getResultLength(IntPtr bidi)
+		public static int ubidi_getResultLength(BiDi.SafeBiDiHandle bidi)
 		{
 			if (BiDiMethods.ubidi_getResultLength == null)
 				BiDiMethods.ubidi_getResultLength = GetMethod<BiDiMethodsContainer.ubidi_getResultLengthDelegate>(IcuCommonLibHandle, "ubidi_getResultLength");
@@ -372,7 +372,7 @@ namespace Icu
 		/// </summary>
 		/// <param name="bidi">The BiDi object</param>
 		/// <returns>The text</returns>
-		public static IntPtr ubidi_getText(IntPtr bidi)
+		public static IntPtr ubidi_getText(BiDi.SafeBiDiHandle bidi)
 		{
 			if (BiDiMethods.ubidi_getText == null)
 				BiDiMethods.ubidi_getText = GetMethod<BiDiMethodsContainer.ubidi_getTextDelegate>(IcuCommonLibHandle, "ubidi_getText");
@@ -385,7 +385,7 @@ namespace Icu
 		/// <param name="bidi">The BiDi object</param>
 		/// <param name="errorCode">The error code</param>
 		/// <returns>The bidi levels</returns>
-		public static IntPtr ubidi_getLevels(IntPtr bidi, out ErrorCode errorCode)
+		public static IntPtr ubidi_getLevels(BiDi.SafeBiDiHandle bidi, out ErrorCode errorCode)
 		{
 			if (BiDiMethods.ubidi_getLevels == null)
 				BiDiMethods.ubidi_getLevels = GetMethod<BiDiMethodsContainer.ubidi_getLevelsDelegate>(IcuCommonLibHandle, "ubidi_getLevels");
@@ -398,7 +398,7 @@ namespace Icu
 		/// <param name="bidi">The BiDi object</param>
 		/// <param name="charIndex">The index of the character in the text.</param>
 		/// <returns>The bidi level</returns>
-		public static byte ubidi_getLevelAt(IntPtr bidi, int charIndex)
+		public static byte ubidi_getLevelAt(BiDi.SafeBiDiHandle bidi, int charIndex)
 		{
 			if (BiDiMethods.ubidi_getLevelAt == null)
 				BiDiMethods.ubidi_getLevelAt = GetMethod<BiDiMethodsContainer.ubidi_getLevelAtDelegate>(IcuCommonLibHandle, "ubidi_getLevelAt");
@@ -410,7 +410,7 @@ namespace Icu
 		/// </summary>
 		/// <param name="bidi">The BiDi object</param>
 		/// <returns>The paragraph level</returns>
-		public static byte ubidi_getParaLevel(IntPtr bidi)
+		public static byte ubidi_getParaLevel(BiDi.SafeBiDiHandle bidi)
 		{
 			if (BiDiMethods.ubidi_getParaLevel == null)
 				BiDiMethods.ubidi_getParaLevel = GetMethod<BiDiMethodsContainer.ubidi_getParaLevelDelegate>(IcuCommonLibHandle, "ubidi_getParaLevel");
@@ -422,7 +422,7 @@ namespace Icu
 		/// </summary>
 		/// <param name="bidi">The BiDi object</param>
 		/// <returns>The number of paragraphs</returns>
-		public static int ubidi_countParagraphs(IntPtr bidi)
+		public static int ubidi_countParagraphs(BiDi.SafeBiDiHandle bidi)
 		{
 			if (BiDiMethods.ubidi_countParagraphs == null)
 				BiDiMethods.ubidi_countParagraphs = GetMethod<BiDiMethodsContainer.ubidi_countParagraphsDelegate>(IcuCommonLibHandle, "ubidi_countParagraphs");
@@ -439,7 +439,7 @@ namespace Icu
 		/// <param name="paraLevel">Will receive the level of the paragraph</param>
 		/// <param name="errorCode">The error code</param>
 		/// <returns>The index of the paragraph containing the specified position</returns>
-		public static int ubidi_getParagraph(IntPtr bidi, int charIndex, out int paraStart, out int paraLimit, out byte paraLevel, out ErrorCode errorCode)
+		public static int ubidi_getParagraph(BiDi.SafeBiDiHandle bidi, int charIndex, out int paraStart, out int paraLimit, out byte paraLevel, out ErrorCode errorCode)
 		{
 			if (BiDiMethods.ubidi_getParagraph == null)
 				BiDiMethods.ubidi_getParagraph = GetMethod<BiDiMethodsContainer.ubidi_getParagraphDelegate>(IcuCommonLibHandle, "ubidi_getParagraph");
@@ -455,7 +455,7 @@ namespace Icu
 		/// <param name="paraLimit">Will receive the limit of the paragraph</param>
 		/// <param name="paraLevel">Will receive the level of the paragraph</param>
 		/// <param name="errorCode">The error code</param>
-		public static void ubidi_getParagraphByIndex(IntPtr bidi, int paraIndex, out int paraStart, out int paraLimit, out byte paraLevel, out ErrorCode errorCode)
+		public static void ubidi_getParagraphByIndex(BiDi.SafeBiDiHandle bidi, int paraIndex, out int paraStart, out int paraLimit, out byte paraLevel, out ErrorCode errorCode)
 		{
 			if (BiDiMethods.ubidi_getParagraphByIndex == null)
 				BiDiMethods.ubidi_getParagraphByIndex = GetMethod<BiDiMethodsContainer.ubidi_getParagraphByIndexDelegate>(IcuCommonLibHandle, "ubidi_getParagraphByIndex");
@@ -468,7 +468,7 @@ namespace Icu
 		/// <param name="bidi">The BiDi object</param>
 		/// <param name="errorCode">The error code</param>
 		/// <returns>The number of runs</returns>
-		public static int ubidi_countRuns(IntPtr bidi, out ErrorCode errorCode)
+		public static int ubidi_countRuns(BiDi.SafeBiDiHandle bidi, out ErrorCode errorCode)
 		{
 			if (BiDiMethods.ubidi_countRuns == null)
 				BiDiMethods.ubidi_countRuns = GetMethod<BiDiMethodsContainer.ubidi_countRunsDelegate>(IcuCommonLibHandle, "ubidi_countRuns");
@@ -482,7 +482,7 @@ namespace Icu
 		/// <param name="visualIndex">The visual position of a character</param>
 		/// <param name="errorCode">The error code</param>
 		/// <returns>The index of this character in the text</returns>
-		public static int ubidi_getLogicalIndex(IntPtr bidi, int visualIndex, out ErrorCode errorCode)
+		public static int ubidi_getLogicalIndex(BiDi.SafeBiDiHandle bidi, int visualIndex, out ErrorCode errorCode)
 		{
 			if (BiDiMethods.ubidi_getLogicalIndex == null)
 				BiDiMethods.ubidi_getLogicalIndex = GetMethod<BiDiMethodsContainer.ubidi_getLogicalIndexDelegate>(IcuCommonLibHandle, "ubidi_getLogicalIndex");
@@ -496,7 +496,7 @@ namespace Icu
 		/// <param name="indexMap">An array of ubidi_getProcessedLength() indexes which will reflect the reordering of the characters;
 		/// if option UBIDI_OPTION_INSERT_MARKS is set, the number of elements allocated in indexMap must be no less than ubidi_getResultLength()</param>
 		/// <param name="errorCode">The error code</param>
-		public static void ubidi_getLogicalMap(IntPtr bidi, [Out, MarshalAs(UnmanagedType.LPArray)]int[] indexMap, out ErrorCode errorCode)
+		public static void ubidi_getLogicalMap(BiDi.SafeBiDiHandle bidi, [Out, MarshalAs(UnmanagedType.LPArray)]int[] indexMap, out ErrorCode errorCode)
 		{
 			if (BiDiMethods.ubidi_getLogicalMap == null)
 				BiDiMethods.ubidi_getLogicalMap = GetMethod<BiDiMethodsContainer.ubidi_getLogicalMapDelegate>(IcuCommonLibHandle, "ubidi_getLogicalMap");
@@ -510,7 +510,7 @@ namespace Icu
 		/// <param name="logicalPosition">A logical position within the source text</param>
 		/// <param name="logicalLimit">Will receive the limit of the corresponding run</param>
 		/// <param name="level">Will receive the level of the corresponding run</param>
-		public static void ubidi_getLogicalRun(IntPtr bidi, int logicalPosition, out int logicalLimit, out byte level)
+		public static void ubidi_getLogicalRun(BiDi.SafeBiDiHandle bidi, int logicalPosition, out int logicalLimit, out byte level)
 		{
 			if (BiDiMethods.ubidi_getLogicalRun == null)
 				BiDiMethods.ubidi_getLogicalRun = GetMethod<BiDiMethodsContainer.ubidi_getLogicalRunDelegate>(IcuCommonLibHandle, "ubidi_getLogicalRun");
@@ -524,7 +524,7 @@ namespace Icu
 		/// <param name="logicalIndex">The index of a character in the tex</param>
 		/// <param name="errorCode">The error code</param>
 		/// <returns>The visual position of this character</returns>
-		public static int ubidi_getVisualIndex(IntPtr bidi, int logicalIndex, out ErrorCode errorCode)
+		public static int ubidi_getVisualIndex(BiDi.SafeBiDiHandle bidi, int logicalIndex, out ErrorCode errorCode)
 		{
 			if (BiDiMethods.ubidi_getVisualIndex == null)
 				BiDiMethods.ubidi_getVisualIndex = GetMethod<BiDiMethodsContainer.ubidi_getVisualIndexDelegate>(IcuCommonLibHandle, "ubidi_getVisualIndex");
@@ -537,7 +537,7 @@ namespace Icu
 		/// <param name="bidi">The BiDi object</param>
 		/// <param name="indexMap">An array of ubidi_getResultLength() indexes which will reflect the reordering of the characters</param>
 		/// <param name="errorCode">The error code</param>
-		public static void ubidi_getVisualMap(IntPtr bidi, [Out, MarshalAs(UnmanagedType.LPArray)]int[] indexMap, out ErrorCode errorCode)
+		public static void ubidi_getVisualMap(BiDi.SafeBiDiHandle bidi, [Out, MarshalAs(UnmanagedType.LPArray)]int[] indexMap, out ErrorCode errorCode)
 		{
 			if (BiDiMethods.ubidi_getVisualMap == null)
 				BiDiMethods.ubidi_getVisualMap = GetMethod<BiDiMethodsContainer.ubidi_getVisualMapDelegate>(IcuCommonLibHandle, "ubidi_getVisualMap");
@@ -552,7 +552,7 @@ namespace Icu
 		/// <param name="logicalStart">The first logical character index in the text</param>
 		/// <param name="length">The number of characters (at least one) in the run</param>
 		/// <returns></returns>
-		public static BiDi.BiDiDirection ubidi_getVisualRun(IntPtr bidi, int runIndex, out int logicalStart, out int length)
+		public static BiDi.BiDiDirection ubidi_getVisualRun(BiDi.SafeBiDiHandle bidi, int runIndex, out int logicalStart, out int length)
 		{
 			if (BiDiMethods.ubidi_getVisualRun == null)
 				BiDiMethods.ubidi_getVisualRun = GetMethod<BiDiMethodsContainer.ubidi_getVisualRunDelegate>(IcuCommonLibHandle, "ubidi_getVisualRun");
@@ -568,7 +568,7 @@ namespace Icu
 		/// <param name="options">Reordering options</param>
 		/// <param name="errorCode">The error code</param>
 		/// <returns>Length of the output string</returns>
-		public static int ubidi_writeReordered(IntPtr bidi, char[] dest, int destSize, ushort options, out ErrorCode errorCode)
+		public static int ubidi_writeReordered(BiDi.SafeBiDiHandle bidi, char[] dest, int destSize, ushort options, out ErrorCode errorCode)
 		{
 			if (BiDiMethods.ubidi_writeReordered == null)
 				BiDiMethods.ubidi_writeReordered = GetMethod<BiDiMethodsContainer.ubidi_writeReorderedDelegate>(IcuCommonLibHandle, "ubidi_writeReordered");